backend/requirements.txt
asgiref==3.5.2
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-allauth==0.51.0
django-bootstrap==0.2.4
django-cors-headers==3.13.0
django-environ==0.8.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
zope.event==4.5.0
zope.interface==5.4.0
